* src/preproc/grn/main.cc: Allow values of `narrow' parameter and friends to be non-integer.
* src/preproc/grn/grn.man: Document it.

* doc/groff.texinfo: Further checking/updating.


* src/preproc/grn/grn.man: Commenting out the -s option -- the
corresponding code doesn't work (yet).

+++ src/preproc/grn/main.cc
@@ -826,16 +826,16 @@ interpret(char *line)
     break;
 
   case 't':			/* thick */
-    thick[2] = defthick[0] * atoi(str2);
+    thick[2] = defthick[0] * atof(str2);
     break;
 
   case 'm':			/* medium */
-    thick[5] = defthick[0] * atoi(str2);
+    thick[5] = defthick[0] * atof(str2);
     break;
 
   case 'n':			/* narrow */
     thick[0] = thick[1] = thick[3] = thick[4] =
-	defthick[0] * atoi(str2);
+	defthick[0] * atof(str2);
     break;
 
   case 'x':			/* x */
